Rename useCases to services as this makes better sense in the context of the UI
