+++ test/services/getCourtCase.integration.test.ts
@@ -4,7 +4,7 @@ import { DataSource } from "typeorm"
 import CourtCase from "../../src/entities/CourtCase"
 import getDataSource from "../../src/lib/getDataSource"
 import { isError } from "../../src/types/Result"
-import getCourtCase from "../../src/useCases/getCourtCase"
+import getCourtCase from "../../src/services/getCourtCase"
 import CourtCaseCase from "../testFixtures/database/data/error_list.json"
 import CourtCaseAho from "../testFixtures/database/data/error_list_aho.json"
 import deleteFromTable from "../testFixtures/database/deleteFromTable"
